aligning test.sh
