+++ scripts/test.sh
@@ -21,46 +21,48 @@ source "${PROG_DIR}/.util/tools.sh"
 source "${PROG_DIR}/.util/print.sh"
 
 function main() {
-  local clean regport dont_create_registry
+  local clean token
   clean="false"
+  token=""
+
+  local regport create_registry
   regport=""
-  dont_create_registry="false"
+  create_registry="true"
 
   while [[ "${#}" != 0 ]]; do
     case "${1}" in
-    --help | -h)
-      shift 1
-      usage
-      exit 0
-      ;;
+      --help|-h)
+        shift 1
+        usage
+        exit 0
+        ;;
+
+      --clean|-c)
+        shift 1
+        clean="true"
+        ;;
+
+      --token|-t)
+        token="${2}"
+        shift 2
+        ;;
 
     --regport | -r)
       regport="${2}"
       shift 2
       ;;
 
-    --dont-create-registry | -dr)
-      dont_create_registry="true"
-      shift 1
-      ;;
+      "")
+        # skip if the argument is empty
+        shift 1
+        ;;
 
-    --clean | -c)
-      shift 1
-      clean="true"
-      ;;
-
-    "")
-      # skip if the argument is empty
-      shift 1
-      ;;
-
-    *)
-      util::print::error "unknown argument \"${1}\""
-      ;;
+      *)
+        util::print::error "unknown argument \"${1}\""
     esac
   done
 
-  tools::install
+  tools::install "${token}"
 
   if [[ "${clean}" == "true" ]]; then
     util::print::title "Cleaning up preexisting stack archives..."
@@ -87,15 +89,21 @@ function main() {
     "${STACK_DIR}/scripts/create.sh"
   fi
 
-  if [[ "${dont_create_registry}" == "true" ]]; then
-    echo "Skipping creating local registry"
+  util::print::title "Setting up local registry"
+
+  if [[ -z "${regport:-}" ]]; then
+    regport="5000"
+  fi
+
+  if [[ "$(curl -s -o /dev/null -w "%{http_code}" localhost:$regport/v2/)" == "200" ]]; then
+    echo "Using existing local registry"
     tests::run $regport ""
   else
+    echo "Creating local registry"
     local local_registry_container_id
     local_registry_container_id=$(docker run -d -p "${regport}:5000" --restart=always --name registry registry:2)
     tests::run $regport $local_registry_container_id
   fi
-
 }
 
 function usage() {
@@ -123,25 +131,26 @@ ${STACK_DIR}/build-java-21/run.oci
 if they exist. Otherwise, first runs create.sh to create them.
 
 OPTIONS
-  --clean  -c  clears contents of stack output directory before running tests
+  --clean         -c  clears contents of stack output directory before running tests
   --regport <port> Local port to use for local registry during tests, defaults to 5000
-  --dont-create-registry -dr skips creating a local registry, uses the existing one
-  --help   -h  prints the command usage
+  --token <token>     Token used to download assets from GitHub (e.g. jam, pack, etc) (optional)
+  --help          -h  prints the command usage
 USAGE
 }
 
 function tools::install() {
+  local token
+  token="${1}"
+
   util::tools::jam::install \
-    --directory "${STACK_DIR}/.bin"
+    --directory "${STACK_DIR}/.bin" \
+    --token "${token}"
 
   util::tools::pack::install \
-    --directory "${STACK_DIR}/.bin"
+    --directory "${STACK_DIR}/.bin" \
+    --token "${token}"
 
   util::tools::skopeo::check
-
-  util::tools::docker::check
-
-  util::tools::git::check
 }
 
 function tests::run() {
@@ -151,24 +160,25 @@ function tests::run() {
   local_registry_port="${1}"
 
   local_registry_container_id="${2}"
+  export REGISTRY_URL="localhost:${local_registry_port}"
 
   pack config experimental true
 
+  export CGO_ENABLED=0
   testout=$(mktemp)
-  pushd "${STACK_DIR}" >/dev/null
-  if GOMAXPROCS="${GOMAXPROCS:-4}" go test -count=1 -timeout 0 ./integration/... -v -run Acceptance --registry-url "localhost:${local_registry_port}" | tee "${testout}"; then
-    util::tools::tests::checkfocus "${testout}"
-    util::print::success "** GO Test Succeeded **"
-  else
-    util::print::error "** GO Test Failed **"
-  fi
+  pushd "${STACK_DIR}" > /dev/null
+    if GOMAXPROCS="${GOMAXPROCS:-4}" go test -count=1 -timeout 0 ./... -v -run Acceptance | tee "${testout}"; then
+      util::tools::tests::checkfocus "${testout}"
+      util::print::success "** GO Test Succeeded **"
+    else
+      util::print::error "** GO Test Failed **"
+    fi
 
   if [[ -n "${local_registry_container_id}" ]]; then
     docker stop $local_registry_container_id
     docker rm $local_registry_container_id
   fi
 
-  popd >/dev/null
+  popd > /dev/null
 }
 
-main "${@:-}"
